+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 1.1
 
 - Added SoundEmitter.java
+- Added ``emitParticles(..., DustOptions dustOptions, ...)``
 
 # Version 1.0